requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
numpy
Pillow
google-genai
python-dotenv
